Move Dino to another folder
